+++ lib/flame_ec2/ec2_api.ex
@@ -81,7 +81,8 @@ defmodule FlameEC2.EC2Api do
         env: env,
         aws_region: config.aws_region,
         s3_bundle_url: config.s3_bundle_url,
-        s3_bundle_compressed?: config.s3_bundle_compressed?
+        s3_bundle_compressed?: config.s3_bundle_compressed?,
+        user_data_pre_script: config.user_data_pre_script
       )
 
     base_params = %{